Script: Factor out Error into a separate module
