+++ libs/chainscript/src/lib.rs
@@ -3,5 +3,9 @@
 #[macro_use]
 mod util;
 
+// TODO: hide implementation of most items exported here
+pub mod error;
 pub mod opcodes;
 pub mod script;
+
+pub use error::Result;
